Refactoring code 2
